apb-6404 [CS] refactor unassigned clients splitting controller - no functional change
