+++ test/controllers/ManageGroupControllerSpec.scala
@@ -107,7 +107,7 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       val groupSummaries = (1 to 3).map(i =>
         GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
@@ -195,7 +195,7 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
 
       val summaries = None
@@ -257,7 +257,7 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render content when filtered clients in session" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       await(sessionCacheRepo.putSession(FILTERED_CLIENTS, displayClients))
       val groupSummaries = (1 to 3).map(i =>
@@ -342,7 +342,7 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render content when filtered access groups and filtered clients in session" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       await(sessionCacheRepo.putSession(FILTERED_CLIENTS, displayClients))
 
@@ -431,7 +431,7 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render correctly the manage groups page" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       expectGetGroupSuccess(groupId, Some(accessGroup))
 
@@ -458,7 +458,7 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render NOT_FOUND when no group is found for this group id" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       expectGetGroupSuccess(groupId, Option.empty[AccessGroup])
 
@@ -487,7 +487,7 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
 
       val groupSummaries = (1 to 3).map(i =>
         GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
@@ -497,7 +497,7 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       expectGetGroupSummarySuccess(arn, summaries)
 
-      implicit val request =
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
         FakeRequest("POST",
           routes.ManageGroupController.submitFilterByGroupName.url)
           .withFormUrlEncodedBody(
@@ -523,12 +523,12 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
 
       val groupSummaries = (1 to 3).map(i =>
         GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
 
-      implicit val request =
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
         FakeRequest("POST",
           routes.ManageGroupController.submitFilterByGroupName.url)
           .withFormUrlEncodedBody(
@@ -552,7 +552,7 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
 
       val groupSummaries = (1 to 3).map(i =>
         GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
@@ -560,7 +560,7 @@ class ManageGroupControllerSpec extends BaseSpec {
         DisplayClient(s"hmrcRef$i", s"name$i", s"HMRC-MTD-IT", ""))
       val summaries = Some((groupSummaries, unassignedClients))
 
-      implicit val request =
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
         FakeRequest("POST",
           routes.ManageGroupController.submitFilterByGroupName.url)
           .withFormUrlEncodedBody(
@@ -598,11 +598,11 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       expectGetGroupSuccess(groupId, Some(accessGroup))
       expectUpdateGroupSuccess(groupId, UpdateAccessGroupRequest(Some("New Group Name"),None,None))
 
-      implicit val request =
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
         FakeRequest("POST",
                     routes.ManageGroupController.submitRenameGroup(groupId).url)
           .withFormUrlEncodedBody("name" -> "New Group Name")
@@ -626,8 +626,8 @@ class ManageGroupControllerSpec extends BaseSpec {
     "redirect when no group is returned for this group id" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      implicit val request =
+      expectIsArnAllowed(allowed = true)
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
         FakeRequest("POST",
                     routes.ManageGroupController.submitRenameGroup(groupId).url)
           .withFormUrlEncodedBody("name" -> "New Group Name")
@@ -656,7 +656,8 @@ class ManageGroupControllerSpec extends BaseSpec {
 
     "render errors when no group name is specified" in {
       //given
-      implicit val request = FakeRequest("POST",
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
+        FakeRequest("POST",
           routes.ManageGroupController.submitRenameGroup(groupId).url)
           .withFormUrlEncodedBody("name" -> "")
           .withHeaders("Authorization" -> s"Bearer whatever")
@@ -664,7 +665,7 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       expectGetGroupSuccess(groupId, Some(accessGroup))
 
       //when
@@ -679,7 +680,7 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render correctly the manage groups page" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       await(sessionCacheRepo.putSession(GROUP_RENAMED_FROM, "Previous Name"))
       expectGetGroupSuccess(groupId, Some(accessGroup))
@@ -715,7 +716,7 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render correctly the DELETE group page" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       expectGetGroupSuccess(groupId, Some(accessGroup))
 
@@ -745,9 +746,9 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render correctly the confirm DELETE group page when 'yes' selected" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
 
-      implicit val request =
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
         FakeRequest("POST",
                     routes.ManageGroupController
                       .submitDeleteGroup(accessGroup._id.toString)
@@ -774,9 +775,9 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render correctly the DASHBOARD group page when 'no' selected" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
 
-      implicit val request =
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
         FakeRequest("POST",
                     routes.ManageGroupController
                       .submitDeleteGroup(accessGroup._id.toString)
@@ -801,7 +802,8 @@ class ManageGroupControllerSpec extends BaseSpec {
 
     "render errors when no answer is specified" in {
       //given
-      implicit val request = FakeRequest("POST",
+      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
+        FakeRequest("POST",
         routes.ManageGroupController.submitDeleteGroup(groupId).url)
         .withFormUrlEncodedBody("answer" -> "")
         .withHeaders("Authorization" -> s"Bearer whatever")
@@ -809,7 +811,7 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       expectGetGroupSuccess(groupId, Some(accessGroup))
 
       //when
@@ -824,7 +826,7 @@ class ManageGroupControllerSpec extends BaseSpec {
     "render correctly" in {
       //given
       expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
+      expectIsArnAllowed(allowed = true)
       await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
       await(sessionCacheRepo.putSession(GROUP_DELETED_NAME, "Rubbish"))
 
@@ -852,351 +854,6 @@ class ManageGroupControllerSpec extends BaseSpec {
         .select("a#returnToDashboard")
         .attr("href") shouldBe routes.ManageGroupController.showManageGroups.url
       html.select(Css.backLink).size() shouldBe 0
-    }
-  }
-
-  s"POST ${routes.ManageGroupController.submitAddUnassignedClients}" should {
-    s"save selected unassigned clients and redirect to ${routes.ManageGroupController.showSelectedUnassignedClients} " +
-      s"when button is Continue" in {
-
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      stubGetClientsOk(arn)(fakeClients)
-
-      implicit val request =
-        FakeRequest("POST", routes.ManageGroupController.submitAddUnassignedClients.url)
-          .withFormUrlEncodedBody(
-            "hasSelectedClients" -> "false",
-            "clients[0]" -> displayClients.head.id,
-            "clients[1]" -> displayClients.last.id,
-            "search" -> "",
-            "filter" -> "",
-            "continue" -> "continue"
-          )
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-
-      val result = controller.submitAddUnassignedClients()(request)
-
-      status(result) shouldBe SEE_OTHER
-
-      redirectLocation(result).get shouldBe s"${routes.ManageGroupController.showSelectedUnassignedClients.url}"
-
-      await(sessionCacheRepo.getFromSession(SELECTED_CLIENTS)) shouldBe Some(Seq(displayClients.head.copy(selected = true),
-        displayClients.last.copy(selected = true)))
-
-    }
-
-    s"save selected unassigned clients and redirect to ${routes.ManageGroupController.showManageGroups}#unassigned-clients " +
-      s"when button is NOT Continue" in {
-
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      stubGetClientsOk(arn)(fakeClients)
-      expectGetGroupSummarySuccess(arn, Some(Seq.empty, displayClients))
-
-
-      implicit val request =
-        FakeRequest("POST", routes.ManageGroupController.submitAddUnassignedClients.url)
-          .withFormUrlEncodedBody(
-            "hasSelectedClients" -> "false",
-            "clients[0]" -> displayClients.head.id,
-            "clients[1]" -> displayClients.last.id,
-            "search" -> "",
-            "filter" -> "VAT",
-            "submitFilter" -> "submitFilter"
-          )
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-
-      val result = controller.submitAddUnassignedClients()(request)
-
-      status(result) shouldBe SEE_OTHER
-
-      redirectLocation(result).get shouldBe s"${routes.ManageGroupController.showManageGroups}#unassigned-clients"
-
-      await(sessionCacheRepo.getFromSession(SELECTED_CLIENTS)) shouldBe Some(Seq(displayClients.head.copy(selected = true),
-        displayClients.last.copy(selected = true)))
-
-    }
-
-    s"present page with errors when form validation fails" in {
-
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      expectGetGroupSummarySuccess(arn, None)
-
-      implicit val request =
-        FakeRequest("POST", routes.ManageGroupController.submitAddUnassignedClients.url)
-          .withFormUrlEncodedBody(
-            "hasSelectedClients" -> "false",
-            "search" -> "",
-            "filter" -> "",
-            "submitFilter" -> "submitFilter"
-          )
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-
-      val result = controller.submitAddUnassignedClients()(request)
-
-      status(result) shouldBe OK
-
-    }
-
-    s"present page with errors when form validation fails and filtered clients exist" in {
-
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      expectGetGroupSummarySuccess(arn, None)
-
-      implicit val request =
-        FakeRequest("POST", routes.ManageGroupController.submitAddUnassignedClients.url)
-          .withFormUrlEncodedBody(
-            "hasSelectedClients" -> "false",
-            "search" -> "",
-            "filter" -> "",
-            "continue" -> "continue"
-          )
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(FILTERED_CLIENTS, displayClients))
-
-      val result = controller.submitAddUnassignedClients()(request)
-
-      status(result) shouldBe OK
-
-      await(sessionCacheRepo.getFromSession(SELECTED_CLIENTS)) shouldBe None
-
-    }
-  }
-
-  s"GET ${routes.ManageGroupController.showSelectedUnassignedClients}" should {
-
-    "redirect if no clients selected are in session" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      expectIsArnAllowed(true)
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-
-      //when
-      val result = controller.showSelectedUnassignedClients(request)
-
-      //then
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result).get shouldBe routes.ManageGroupController.showManageGroups.url
-    }
-
-    "render correctly the selected unassigned clients page" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-
-      //when
-      val result = controller.showSelectedUnassignedClients(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title() shouldBe "Review selected clients - Agent services account - GOV.UK"
-      html.select(H1).text() shouldBe "You have selected 3 clients"
-      html.select(Css.tableWithId("sortable-table")).select("tbody tr").size() shouldBe 3
-      //and the back link should go to the unassigned clients tab
-      html.select(Css.backLink).attr("href") shouldBe "/agent-permissions/manage-access-groups#unassigned-clients"
-
-      html.select("form .govuk-fieldset__legend").text() shouldBe "Do you need to add or remove selected clients?"
-      val answerRadios = html.select(Css.radioButtonsField("answer"))
-      answerRadios
-        .select("label[for=true]")
-        .text() shouldBe "Yes, add or remove clients"
-      answerRadios
-        .select("label[for=false]")
-        .text() shouldBe "No, continue to next section"
-      html.select(Css.submitButton).text() shouldBe "Save and continue"
-
-    }
-  }
-
-  s"GET ${routes.ManageGroupController.showSelectGroupsForSelectedUnassignedClients}" should {
-
-    "render correctly the select groups for unassigned clients page" in {
-      //given
-      val groupSummaries = (1 to 3).map(i =>
-        GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      expectGetGroupSummarySuccess(arn, Some((groupSummaries, Seq.empty[DisplayClient])))
-
-      //when
-      val result = controller.showSelectGroupsForSelectedUnassignedClients(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title() shouldBe "Which access groups would you like to add the selected clients to? - Agent services account - GOV.UK"
-      html.select(H1).text() shouldBe "Which access groups would you like to add the selected clients to?"
-      //and the back link should go to the unassigned clients tab
-      html.select(Css.backLink).attr("href") shouldBe routes.ManageGroupController.showSelectedUnassignedClients.url
-
-    }
-  }
-
-  s"POST ${routes.ManageGroupController.submitSelectGroupsForSelectedUnassignedClients}" should{
-
-    "redirect to create group if CREATE NEW is selected" in {
-      //given
-      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
-        FakeRequest("POST", routes.ManageGroupController.submitSelectGroupsForSelectedUnassignedClients.url)
-          .withFormUrlEncodedBody("createNew" -> "true")
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-
-      //when
-      val result = controller.submitSelectGroupsForSelectedUnassignedClients(request)
-
-      //then
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result).get shouldBe routes.CreateGroupController.showGroupName.url
-
-    }
-
-    "redirect to confirmation page when existing groups are selected to assign the selected clients to" in {
-      //given
-      val groupSummaries = (1 to 3).map(i => GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
-
-      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
-        FakeRequest("POST", routes.ManageGroupController.submitSelectGroupsForSelectedUnassignedClients.url)
-          .withFormUrlEncodedBody("groups[0]" -> "12412312")
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      expectGetGroupSummarySuccess(arn, Some((groupSummaries, Seq.empty[DisplayClient])))
-
-      //when
-      val result = controller.submitSelectGroupsForSelectedUnassignedClients(request)
-
-      //then
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result).get shouldBe routes.ManageGroupController.showConfirmClientsAddedToGroups.url
-
-    }
-
-    "show errors when nothing selected" in {
-      //given
-      val groupSummaries = (1 to 3).map(i =>
-        GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
-
-      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
-        FakeRequest("POST", routes.ManageGroupController.submitSelectGroupsForSelectedUnassignedClients.url)
-          .withFormUrlEncodedBody()
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      expectGetGroupSummarySuccess(arn, Some((groupSummaries, Seq.empty[DisplayClient])))
-
-      //when
-      val result = controller.submitSelectGroupsForSelectedUnassignedClients(request)
-
-      //then
-      status(result) shouldBe OK
-      //and should show errors
-      val html = Jsoup.parse(contentAsString(result))
-      html.select(Css.errorSummaryForField("field-wrapper")).text() shouldBe "You must select an access group or add a new group"
-      html.select(Css.errorForField("field-wrapper")).text() shouldBe "You must select an access group or add a new group"
-
-
-    }
-
-    "show errors when both createNew and existing groups are selected" in {
-      //given
-      val groupSummaries = (1 to 3).map(i =>
-        GroupSummary(s"groupId$i", s"name $i", i * 3, i * 4))
-
-      implicit val request: FakeRequest[AnyContentAsFormUrlEncoded] =
-        FakeRequest("POST", routes.ManageGroupController.submitSelectGroupsForSelectedUnassignedClients.url)
-          .withFormUrlEncodedBody("createNew" -> "true","groups[0]" -> "12412312")
-          .withSession(SessionKeys.sessionId -> "session-x")
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-      expectGetGroupSummarySuccess(arn, Some((groupSummaries, Seq.empty[DisplayClient])))
-
-      //when
-      val result = controller.submitSelectGroupsForSelectedUnassignedClients(request)
-
-      //then
-      status(result) shouldBe OK
-      //and should show errors
-      val html = Jsoup.parse(contentAsString(result))
-      html.select(Css.errorSummaryForField("field-wrapper")).text() shouldBe "You cannot add to existing groups at the same time as creating a new group"
-      html.select(Css.errorForField("field-wrapper")).text() shouldBe "You cannot add to existing groups at the same time as creating a new group"
-
-
-    }
-  }
-
-  s"GET ${routes.ManageGroupController.showConfirmClientsAddedToGroups}" should {
-
-    "render correctly the select groups for unassigned clients page" in {
-      //given
-      val groups = Seq("South West", "London")
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(GROUPS_FOR_UNASSIGNED_CLIENTS, groups))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-
-      //when
-      val result = controller.showConfirmClientsAddedToGroups(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title() shouldBe "Clients added to access groups - Agent services account - GOV.UK"
-      html.select(H1).text() shouldBe "Clients added to access groups"
-      val paragraphs = html.select(Css.paragraphs)
-      paragraphs.get(0).text() shouldBe "You have added these clients to the following groups:"
-      val listItems = html.select("ul.govuk-list li.govuk-list--item")
-      listItems.size() shouldBe groups.size
-      listItems.get(0).text shouldBe groups(0)
-      listItems.get(1).text shouldBe groups(1)
-      //and the back link should go to the unassigned clients tab
-      html.select(Css.backLink).size() shouldBe 0
-
-    }
-
-    "redirect when no group names in session" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectIsArnAllowed(true)
-
-      //when
-      val result = controller.showConfirmClientsAddedToGroups(request)
-
-      //then
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result).get shouldBe routes.ManageGroupController.showSelectGroupsForSelectedUnassignedClients.url
-
     }
   }
